+++ src/GUI/CItemSlotsBar.cpp
@@ -149,7 +149,7 @@ void CItemSlotsBar::SetSlotBackground(const std::string& bg)
     }
 }
 
-void CItemSlotsBar::SetSlotItem(unsigned int slot, CItem* item, bool updatePlayer)
+void CItemSlotsBar::SetSlotItem(size_t slot, CItem* item, bool updatePlayer)
 {
     if (slot < ITEM_SLOTS_COUNT && mSlot[slot])
         mSlot[slot]->SetSelectedItem(item);
